corrections and tweaks
git-svn-id: https://svn.r-project.org/R-packages/trunk@7755 694ef91d-65df-0310-b7bb-92e67a308ead

+++ XML/src/DocParse.c
@@ -199,7 +199,7 @@ RS_XML(ParseTree)(USER_OBJECT_ fileName, USER_OBJECT_ converterFunctions,
 
   if(asTextBuffer) {
       doc = useHTML ? htmlParseDoc(CHAR_TO_XMLCHAR(name), encoding) : 
-	              xmlReadMemory(name, strlen(name), NULL, encoding, parserOptions) ;
+	  xmlReadMemory(name, (int)strlen(name), NULL, encoding, parserOptions) ;
                 	  /* xmlParseMemory(name, strlen(name)) */ 
 
       if(doc != NULL)
