+++ XML/src/XMLEventParse.c
@@ -80,7 +80,7 @@ createSAX2AttributesList(const xmlChar **attributes, int nb_attributes, int nb_d
       char *tmp;
       int len;
 
-      len = (ptr[4] - ptr[3] + 1);
+      len = (int)(ptr[4] - ptr[3] + 1);
       tmp = malloc(sizeof(char) * len);
       if(!tmp) {
          PROBLEM "Cannot allocate space for attribute of length %d", (int) (ptr[4] - ptr[3] + 2)
@@ -192,7 +192,7 @@ RS_XML_readConnectionInput(void *context, char *buffer, int len)
    if(GET_LENGTH(tmp)) {
 
       str = CHAR_DEREF(STRING_ELT(tmp, 0));
-      n = strlen(str);
+      n = (int)strlen(str);
 
 
       if(n != 0) { /* Just add a new line and do it again. */
@@ -343,7 +343,7 @@ getPropertyValue(const xmlChar **ptr)
   int len;
   char *tmp;
 
-      len = (ptr[4] - ptr[3] + 1);
+  len = (int)(ptr[4] - ptr[3] + 1);
       tmp = malloc(sizeof(char) * len);
       if(!tmp) {
          PROBLEM "Cannot allocate space for attribute of length %d", (int) (ptr[4] - ptr[3] + 2)
@@ -792,7 +792,7 @@ do_getEntityHandler(void *userData, const xmlChar *name, const char * r_funName)
 	    ans->name = xmlStrdup(name);
 	    ans->orig = NULL; // xmlStrdup(CHAR_TO_XMLCHAR(value));
 	    ans->content = xmlStrdup(CHAR_TO_XMLCHAR(value));	    
-	    ans->length = strlen(value);
+	    ans->length = (int)strlen(value);
 #ifndef NO_CHECKED_ENTITY_FIELD
 	    ans->checked = 1;
 #endif
